requirements.txt
flask==3.1.*
gunicorn==23.*
pillow==11.*
beautifulsoup4==4.*
requests==2.*
